new setup for efficient qs
